# backend/coze_plugin_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "顺发 Coze Plugin API",
    "description": "顺发AI洞察者后端，供Coze工作流调用",
    "version": "1.0.0"
  },
  "paths": {
    "/api/coze/get_hot_topics": {
      "get": {
        "summary": "Get Hot Topics",
        "description": "Return today's top-scored pending hot topics for the daily push.",
        "operationId": "get_hot_topics_api_coze_get_hot_topics_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 3,
              "title": "Limit"
            }
          },
          {
            "name": "X-Coze-Plugin-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Coze-Plugin-Token"
            }
          },
          {
            "name": "X-Feishu-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "首选的飞书用户ID；若未注入，后端会尝试兼容其他用户标识头并降级到匿名Coze用户"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetHotTopicsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CozePluginToken": []
          }
        ]
      }
    },
    "/api/coze/quick_generate": {
      "post": {
        "summary": "Coze Quick Generate",
        "description": "Quick mode: generate content from hot topic + angle in ~30s. Stateless.",
        "operationId": "coze_quick_generate_api_coze_quick_generate_post",
        "parameters": [
          {
            "name": "X-Coze-Plugin-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Coze-Plugin-Token"
            }
          },
          {
            "name": "X-Feishu-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "首选的飞书用户ID；若未注入，后端会尝试兼容其他用户标识头并降级到匿名Coze用户"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickGeneratePluginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickGeneratePluginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CozePluginToken": []
          }
        ]
      }
    },
    "/api/coze/start_deep_mode": {
      "post": {
        "summary": "Start Deep Mode",
        "description": "Create a deep mode CheckIn session. Returns checkin_id for subsequent messages.",
        "operationId": "start_deep_mode_api_coze_start_deep_mode_post",
        "parameters": [
          {
            "name": "X-Coze-Plugin-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Coze-Plugin-Token"
            }
          },
          {
            "name": "X-Feishu-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "首选的飞书用户ID；若未注入，后端会尝试兼容其他用户标识头并降级到匿名Coze用户"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartDeepModeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartDeepModeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CozePluginToken": []
          }
        ]
      }
    },
    "/api/coze/deep_mode_message": {
      "post": {
        "summary": "Deep Mode Message",
        "description": "Send a message in the deep mode discussion flow.",
        "operationId": "deep_mode_message_api_coze_deep_mode_message_post",
        "parameters": [
          {
            "name": "X-Coze-Plugin-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Coze-Plugin-Token"
            }
          },
          {
            "name": "X-Feishu-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "首选的飞书用户ID；若未注入，后端会尝试兼容其他用户标识头并降级到匿名Coze用户"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeepModeMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeepModeMessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CozePluginToken": []
          }
        ]
      }
    },
    "/api/coze/confirm_and_publish": {
      "post": {
        "summary": "Confirm And Publish",
        "description": "Confirm content and publish. Updates streak and points.",
        "operationId": "confirm_and_publish_api_coze_confirm_and_publish_post",
        "parameters": [
          {
            "name": "X-Coze-Plugin-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Coze-Plugin-Token"
            }
          },
          {
            "name": "X-Feishu-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "首选的飞书用户ID；若未注入，后端会尝试兼容其他用户标识头并降级到匿名Coze用户"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/app__routers__coze_plugin__ConfirmPublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmPublishResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CozePluginToken": []
          }
        ]
      }
    },
    "/api/coze/user_stats": {
      "get": {
        "summary": "Get User Stats",
        "description": "Return user streak, points, level, and today's completion status.",
        "operationId": "get_user_stats_api_coze_user_stats_get",
        "parameters": [
          {
            "name": "X-Coze-Plugin-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Coze-Plugin-Token"
            }
          },
          {
            "name": "X-Feishu-User-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "首选的飞书用户ID；若未注入，后端会尝试兼容其他用户标识头并降级到匿名Coze用户"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "CozePluginToken": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConfirmPublishResponse": {
        "properties": {
          "streak": {
            "type": "integer",
            "title": "Streak"
          },
          "points_earned": {
            "type": "integer",
            "title": "Points Earned"
          },
          "total_points": {
            "type": "integer",
            "title": "Total Points"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "newly_unlocked": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Newly Unlocked",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "streak",
          "points_earned",
          "total_points",
          "level",
          "message"
        ],
        "title": "ConfirmPublishResponse"
      },
      "DeepModeMessageRequest": {
        "properties": {
          "checkin_id": {
            "type": "integer",
            "title": "Checkin Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "angle": {
            "type": "string",
            "title": "Angle",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "checkin_id",
          "message"
        ],
        "title": "DeepModeMessageRequest"
      },
      "DeepModeMessageResponse": {
        "properties": {
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "draft": {
            "type": "string",
            "title": "Draft",
            "default": ""
          },
          "has_draft": {
            "type": "boolean",
            "title": "Has Draft",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "reply",
          "status"
        ],
        "title": "DeepModeMessageResponse"
      },
      "GetHotTopicsResponse": {
        "properties": {
          "topics": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Topics"
          },
          "date": {
            "type": "string",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "topics",
          "date"
        ],
        "title": "GetHotTopicsResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuickGeneratePluginRequest": {
        "properties": {
          "hot_topic": {
            "type": "string",
            "title": "Hot Topic"
          },
          "angle": {
            "type": "string",
            "title": "Angle"
          },
          "platform": {
            "type": "string",
            "title": "Platform",
            "default": "xiaohongshu"
          }
        },
        "type": "object",
        "required": [
          "hot_topic",
          "angle"
        ],
        "title": "QuickGeneratePluginRequest"
      },
      "QuickGeneratePluginResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          }
        },
        "type": "object",
        "required": [
          "content",
          "platform",
          "char_count"
        ],
        "title": "QuickGeneratePluginResponse"
      },
...
      "StartDeepModeRequest": {
        "properties": {
          "hot_topic": {
            "type": "string",
            "title": "Hot Topic"
          },
          "angle": {
            "type": "string",
            "title": "Angle"
          }
        },
        "type": "object",
        "required": [
          "hot_topic",
          "angle"
        ],
        "title": "StartDeepModeRequest"
      },
      "StartDeepModeResponse": {
        "properties": {
          "checkin_id": {
            "type": "integer",
            "title": "Checkin Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "opening_message": {
            "type": "string",
            "title": "Opening Message"
          }
        },
        "type": "object",
        "required": [
          "checkin_id",
          "status",
          "opening_message"
        ],
        "title": "StartDeepModeResponse"
      },
...
      "UserStatsResponse": {
        "properties": {
          "streak": {
            "type": "integer",
            "title": "Streak"
          },
          "longest_streak": {
            "type": "integer",
            "title": "Longest Streak"
          },
          "points": {
            "type": "integer",
            "title": "Points"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "diamonds": {
            "type": "integer",
            "title": "Diamonds"
          },
          "today_completed": {
            "type": "boolean",
            "title": "Today Completed"
          },
          "achievements": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Achievements",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "streak",
          "longest_streak",
          "points",
          "level",
          "diamonds",
          "today_completed"
        ],
        "title": "UserStatsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__routers__coze_plugin__ConfirmPublishRequest": {
        "properties": {
          "checkin_id": {
            "type": "integer",
            "title": "Checkin Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "checkin_id",
          "content"
        ],
        "title": "ConfirmPublishRequest"
      },
...
    },
    "securitySchemes": {
      "CozePluginToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Coze-Plugin-Token"
      }
    }
  }
}